requirements.txt
42cc-pystyle==0.0.16
appdirs==1.4.3
Django==1.6.7
django-floppyforms==1.7.0
flake8==2.6.2
mccabe==0.5.3
model-mommy==1.3.2
olefile==0.44
packaging==16.8
Pillow==4.1.1
pycodestyle==2.0.0
pyflakes==1.2.3
pyparsing==2.2.0
six==1.10.0
South==1.0.2
